Move routes to network folder
